+++ assets/scripts/puzzle-ui.js
@@ -827,6 +827,7 @@ export class PuzzleUi {
                 if (this.config.stats.ms_until_due <= 0) {
                     console.log('Countdown done');
                     clearInterval(this.countdown_interval);
+                    this.countdown_interval = null;
                     this.request_data();
                 }
             }, 1000);