+++ src/main.rs
@@ -13,6 +13,9 @@ mod util;
 use std::env;
 use std::error::Error;
 use std::net::SocketAddr;
+use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::time::Duration;
 use futures::TryFutureExt;
 use tokio_cron_scheduler::{Job, JobScheduler};
 
@@ -51,34 +54,45 @@ async fn main() -> Result<(), Box<dyn Error>> {
     };
 
     // Initialise puzzle database in background if necessary.
-    tokio::spawn(lichess::init_db(puzzle_db.clone())
+    let cancel_import = Arc::new(AtomicBool::new(false));
+    let import_done = Arc::new(AtomicBool::new(false));
+    tokio::spawn(lichess::init_db(puzzle_db.clone(), cancel_import.clone())
         .and_then({
+            let import_done = import_done.clone();
+            let cancel_import = cancel_import.clone();
             let app_config = app_config.clone();
             let url = url.clone();
             move |data_imported| {
-                if data_imported {
+                if data_imported && !cancel_import.load(Ordering::Relaxed) {
                     log::info!("Puzzle database import complete");
                     log::info!("Reminder: the application is now listening on {}:{}",
                                app_config.bind_interface, app_config.bind_port);
                     log::info!("Access it at {url}");
                 }
+                import_done.store(true, Ordering::Relaxed);
                 async { Ok(()) }
             }
         })
-        .or_else(|e| {
-            log::info!("{e}");
-            async { Err(e) }
+        .or_else({
+            let import_done = import_done.clone();
+            move |e| {
+                log::info!("{e}");
+                import_done.store(true, Ordering::Relaxed);
+                async { Err(e) }
+            }
         }));
 
     // Create application routes.
-    let app_state = AppState::new(app_config.clone(), puzzle_db);
+    let app_state = AppState::new(app_config.clone(), puzzle_db.clone());
     let app = controllers::routes(app_state.clone())
         .nest_service("/api", api::routes(app_state))
         .nest_service(assets::STATIC_ASSETS_PATH, assets::routes());
 
     // Start web server.
     let socket_addr = SocketAddr::from((app_config.bind_interface, app_config.bind_port));
-    let server_task = tokio::spawn(axum::Server::bind(&socket_addr).serve(app.into_make_service()));
+    let server_task = tokio::spawn(hyper::Server::bind(&socket_addr)
+        .serve(app.into_make_service())
+        .with_graceful_shutdown(shutdown_signal(puzzle_db, import_done, cancel_import)));
 
     // Print server address.
     log::info!("The application is now listening on {}:{}", app_config.bind_interface, app_config.bind_port);
@@ -122,4 +136,31 @@ async fn try_run_backup(app_config: AppConfig, db: PuzzleDatabase) {
     if let Err(e) = app::backup::run_backup(app_config, db).await {
         log::error!("Error when backing up database: {e}");
     }
+}
+
+/// Returns a future that awaits a shutdown signal and performs graceful shutdown.
+async fn shutdown_signal(db: PuzzleDatabase, import_done: Arc<AtomicBool>,
+    cancel_import: Arc<AtomicBool>)
+{
+    // Await ctrl-c signal.
+    tokio::signal::ctrl_c()
+        .await
+        .expect("Failed to await ctrl-c signal");
+
+    log::info!("Termination requested, closing database...");
+
+    // If the lichess puzzle database is still importing, stop it.
+    if !import_done.load(Ordering::Relaxed) {
+        cancel_import.store(true, Ordering::Relaxed);
+
+        // Wait for the import task to stop.
+        while !import_done.load(Ordering::Relaxed) {
+            log::info!("Waiting for lichess puzzle db import to stop...");
+            tokio::time::sleep(Duration::from_secs(1)).await;
+        }
+    }
+
+    // Commit and close database.
+    db.close().await
+        .expect("Failed to close database connection");
 }