+++ CHANGELOG.md
@@ -6,9 +6,15 @@ Added:
   automatically be set to failed if the hint is used, but the user can always override it on the
   'Puzzle complete' panel and submit a positive response if desired.
 
+Changed:
+* Ctrl-c now causes a graceful shutdown and commits the database journal before closing it.
+
 Fixed:
-* A small fix to the scrolwheel behavior on the puzzle board to prevent the page from scrolling
+* A small fix to the scroll wheel behavior on the puzzle board to prevent the page from scrolling
   if the mousewheel is over the board.
+* Fix countdown timer to next review not starting sometimes.
+* The 'next review due' time on the stats screen no longer says 'never' when there are no reviews
+  due for the rest of the day.
 
 [1.9.0] - 2023-10-26